Moved all hooks to FQLHooks / comments
