+++ src/typedefs.js
@@ -1,0 +1,3 @@
+/**
+ * @external {jQuery} https://api.jquery.com/
+ */
